feat(engine): Add Position2D trait + move Pos2 to layout module
