+++ mindy-engine/src/layout/pos2.rs
@@ -1,4 +1,5 @@
 use serde::{Deserialize, Serialize};
+use crate::layout::Position2D;
 
 #[derive(Debug, Clone, Default, PartialEq, Deserialize, Serialize)]
 pub struct Pos2 {
@@ -12,18 +13,22 @@ impl Pos2 {
     pub fn new(x: f32, y: f32) -> Self {
         Self { x, y }
     }
+}
 
-    pub fn subtract(&self, other: &Self) -> Self {
-        Self {
-            x: self.x - other.x,
-            y: self.y - other.y,
-        }
+impl Position2D for Pos2 {
+    fn x(&self) -> f32 {
+        self.x
     }
 
-    pub fn add(&self, other: &Self) -> Self {
-        Self {
-            x: self.x + other.x,
-            y: self.y + other.y,
-        }
+    fn y(&self) -> f32 {
+        self.y
+    }
+
+    fn x_mut(&mut self) -> &mut f32 {
+        &mut self.x
+    }
+
+    fn y_mut(&mut self) -> &mut f32 {
+        &mut self.y
     }
 }
