+++ mindy-engine/src/lib.rs
@@ -2,6 +2,7 @@ pub mod link;
 pub mod mindmap;
 pub mod node;
 pub mod utils;
+pub mod layout;
 
 pub fn add(left: u64, right: u64) -> u64 {
     left + right
