+++ mindy-engine/src/mindmap/mod.rs
@@ -1,11 +1,13 @@
 use crate::mindmap::metadata::MindmapMetadata;
 use crate::mindmap::r#type::MindmapType;
 use crate::node::Node;
-use crate::utils::pos2::Pos2;
+use crate::layout::pos2::Pos2;
 use crate::utils::size::Size;
 use serde::{Deserialize, Serialize};
 use std::error::Error;
+use crate::layout::Position2D;
 use crate::mindmap::layout::LayoutEngine;
+
 
 pub mod metadata;
 pub mod style;
@@ -147,7 +149,7 @@ impl Mindmap {
             "get_node_bounding_box - min_x: {:?}, min_y: {:?}, width: {:?}, height: {:?}",
             min_x, min_y, width, height
         );
-        self.position = Some(Pos2::new(min_x, min_y).subtract(&extra_offset));
+        self.position = Some(Pos2::new(min_x, min_y).subtract(&extra_offset).clone());
         self.size = Some(Size { width, height });
         self
     }
