+++ mindy-engine/src/utils/mod.rs
@@ -1,4 +1,3 @@
-pub mod pos2;
 pub mod rgb;
 pub mod size;
 pub mod throttler;
