+++ mindy-html/src/mindmap/mod.rs
@@ -6,7 +6,7 @@ use base64::Engine;
 use dioxus::logger::tracing;
 use dioxus::prelude::*;
 use mindy_engine::mindmap::Mindmap;
-use mindy_engine::utils::pos2::Pos2;
+use mindy_engine::layout::pos2::Pos2;
 use mindy_engine::utils::size::Size;
 
 pub(crate) static MINDMAP: GlobalSignal<Mindmap> = GlobalSignal::new(|| Mindmap::default());
